refactor(site): split agents limits tab sections
