+++ site/src/pages/AgentsPage/LimitsTab/index.ts
@@ -1,0 +1,1 @@
+export { LimitsTab } from "./LimitsTab";
